requirements.txt
streamlit
openai
faiss-cpu
PyMuPDF
tiktoken
python-dotenv
